[pthread][libc] move 'clock_time' to libc
