--- components/libc/time/SConscript
@@ -1,12 +1,0 @@
-# RT-Thread building script for component
-
-from building import *
-
-cwd = GetCurrentDir()
-src = Glob('*.c') + Glob('*.cpp')
-CPPPATH = [cwd]
-
-group = DefineGroup('libc', src, 
-    depend = ['RT_USING_PTHREADS'], CPPPATH = CPPPATH)
-
-Return('group')
